backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
pydantic>=2.0
pydantic-settings>=2.0
python-multipart>=0.0.9
openai>=1.0.0
opencv-python>=4.10.0
Pillow>=10.0
numpy>=1.26.0
